src/main/java/com/api/investmanager/adapters/input/provider/UserProviderAuth.java
package com.api.investmanager.adapters.input.provider;

import com.api.investmanager.core.application.port.input.user.AuthenticatedUserProvider;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Component;

@Component
public class UserProviderAuth implements AuthenticatedUserProvider {

    private Authentication getAuthentication() {
        return SecurityContextHolder.getContext().getAuthentication();
    }

    @Override
    public String getUserId() {
        Authentication authentication = getAuthentication();
        if (authentication != null && authentication.isAuthenticated()) {
            return (String) authentication.getPrincipal();
        }
        return null;
    }
}
